[move-only] Move Receiver and test code into kind::unified::address.
